text_toner_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pymongo==4.6.0
python-dotenv==1.0.0
transformers==4.36.0
torch
numpy>=1.26.0
scikit-learn==1.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
